refactor: split price api service
